pyproject.toml
[build-system]
requires = [
  "setuptools>=61.0",
  "versioneer[toml]",
  'tomli; python_version < "3.11"',
 ]
build-backend = "setuptools.build_meta"

[project]
name = "CreateCloudMap"
dynamic = ["version"]
authors = [
  { name="Joachim Herb", email="Joachim.Herb@gmx.de" },
]
description = "Downloads a cloud map for xplanet from https://clouds.matteason.co.uk/"
readme = "README.rst"
license = {file = "LICENSE"}
requires-python = ">=3.8"
dependencies = [
  'requests',
  'python-dateutil'
]
classifiers = [
        "Programming Language :: Python",
        "Programming Language :: Python :: 3",
        "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
        "Operating System :: OS Independent",
        "Development Status :: 4 - Beta",
        "Environment :: Console",
        "Intended Audience :: End Users/Desktop",
        "Topic :: Desktop Environment",
        "Topic :: Multimedia :: Graphics",
        "Topic :: Scientific/Engineering :: Atmospheric Science",
        "Topic :: Scientific/Engineering :: GIS",
        "Topic :: Scientific/Engineering :: Visualization",
        "Topic :: Utilities",
]

[project.scripts]
create_map = "cloudmap.create_map:main"

[project.urls]
Homepage = "https://github.com/jmozmoz/cloudmap"
Issues = "https://github.com/jmozmoz/cloudmap/issues"

[tool.versioneer]
VCS = "git"
style = "pep440"
versionfile_source = "cloudmap/_version.py"
versionfile_build = "cloudmap/_version.py"
tag_prefix = ""
parentdir_prefix = "CreateCloudMap-"